requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
reportlab==4.4.5
requests==2.32.5
sqlparse==0.5.4
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.6.1
whitenoise==6.11.0
